server/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5
django-debug-toolbar==2.2
python-telegram-bot==12.8